% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{fit_polynomial}
\alias{fit_polynomial}
\title{Fit polynomial regression}
\usage{
fit_polynomial(data, degree)
}
\arguments{
\item{data}{Training dataset}

\item{degree}{Polynomial degree}
}
\value{
A fitted lm object.
}
\description{
Fits y ~ poly(x, degree, raw = TRUE) using least squares.
}
\examples{
set.seed(1)
f <- true_function("wiggly")
train <- generate_data(n = 80, sigma = 1, f = f)
mod <- fit_polynomial(train, degree = 5)
summary(mod)
}
