% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{generate_data}
\alias{generate_data}
\title{Generate regression data}
\usage{
generate_data(n, sigma, f)
}
\arguments{
\item{n}{Sample size}

\item{sigma}{Noise level}

\item{f}{True function}
}
\value{
data.frame with x and y
}
\description{
Generate regression data
}
