% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{monte_carlo_bias_variance}
\alias{monte_carlo_bias_variance}
\title{Monte Carlo simulation for Bias-Variance Tradeoff}
\usage{
monte_carlo_bias_variance(B, n, sigma, degree, f)
}
\arguments{
\item{B}{Number of Monte Carlo replications.}

\item{n}{Sample size for each training dataset.}

\item{sigma}{Noise level (standard deviation of the Gaussian noise).}

\item{degree}{Polynomial degree to fit.}

\item{f}{The true data-generating function.}
}
\value{
A list containing the estimated \code{bias2} and \code{variance}.
}
\description{
Estimates bias squared and variance for a given polynomial degree using Monte Carlo simulation.
}
